MDL-66413 mod_feedback: Limit faulty use of captchas
If google captcha was not configured properly, teachers should not be
able to use captcha as a feedback item - this would result in students
unable to complete the feedback as they cannot complete the captcha.

Added an adminsetting to disable the use of captchas as feedback items
independently.

Under which conditions can captcha feedback be used?
 If the Admin configured captcha credentials and the adminsetting is enabled
 (default), Teachers can choose to insert the feedback item captcha.

Negative conditions:
1. Scenario: No captcha credentials were configured. Teachers cannot choose to
 insert the feedback item Captcha
2. Scenario: Captcha credentials are configured. Admin disabled the
 allowgcaptcha setting.
 Teachers cannot choose to insert feedback item Captcha